+++ client/src/components/_common/Section/SectionMessages.test.js
@@ -5,7 +5,6 @@ import { Provider } from 'react-redux';
 import { Alert } from 'reactstrap';
 
 import SectionMessages from './SectionMessages';
-import * as MESSAGES from '../../../constants/messages';
 
 const mockStore = configureStore();
 const store = mockStore({});
@@ -28,33 +27,14 @@ describe('SectionMessages', () => {
     });
   });
 
-  describe('weclome message', () => {
-    it('renders known intro message', () => {
-      const { getByText } = render(
-        <Provider store={store}>
-          <SectionMessages messageComponentName="DASHBOARD" />
-        </Provider>
-      );
-      expect(getByText(MESSAGES['DASHBOARD'])).not.toEqual(null);
-    });
-
-    it('renders known intro message but with custom message', () => {
+  describe('custom messages', () => {
+    it('renders custom intro message', () => {
       const { getByText, queryByText } = render(
         <Provider store={store}>
           <SectionMessages
             messageComponentName="DASHBOARD"
             introMessageText="Hello"
           />
-        </Provider>
-      );
-      expect(queryByText(MESSAGES['DASHBOARD'])).toEqual(null);
-      expect(getByText('Hello')).not.toEqual(null);
-    });
-
-    it('renders custom intro message', () => {
-      const { getByText } = render(
-        <Provider store={store}>
-          <SectionMessages introMessageText="Hello" />
         </Provider>
       );
       expect(getByText('Hello')).not.toEqual(null);