+++ server/portal/settings/settings_custom.example.py
@@ -258,6 +258,22 @@ _WORKBENCH_SETTINGS = {
     "hideApps": False,
     "hideDataFiles": False,
     "hideAllocations": False,
+    "showSubmissions": False,
+    "hideManageAccount": False,
+    "hasUserGuide": True,
     "onboardingCompleteRedirect": '/workbench/',
-    "noPHISystem": ""
+    "noPHISystem": "",
+    "customDashboardSection": {
+        "header": "My Account",
+        "links": [
+            {
+                "href": 'https://utexas.edu',
+                "text": 'Update Profile and Email Address',
+            },
+            {
+                "href": 'https://utexas.edu',
+                "text": 'Change Password',
+            }
+        ]
+    }
 }